Update PDF_ScraperV1
In this updated script, I added a print(message) line in the log_to_file function. This will print the log message to the terminal in addition to writing it to the log file. Now you should be able to see the log messages in your terminal while the script is running, as well as in the log file after it has finished running.